Check if Firebase reference path is a valid one before deletion.

+++ mapswipe_workers/mapswipe_workers/firebase_to_postgres/archive_project.py
@@ -1,19 +1,19 @@
 """
 Archive a project.
 """
-
+import re
 from typing import Iterable
 
 from firebase_admin import exceptions
 
 from mapswipe_workers import auth
-from mapswipe_workers.definitions import logger
+from mapswipe_workers.definitions import logger, CustomError
 
 
 def chunks(data: list, size: int = 250) -> Iterable[list]:
     """Yield successive n-sized chunks from list."""
     for i in range(0, len(data), size):
-        yield data[i : i + size]
+        yield data[i : i + size]  # noqa E203
 
 
 def archive_project(project_ids: list) -> None:
@@ -28,6 +28,11 @@ def archive_project(project_ids: list) -> None:
 
         fb_db = auth.firebaseDB()
         ref = fb_db.reference(f"v2/results/{project_id}")
+        if not re.match(r"/v2/\w+/[-a-zA-Z0-9]+", ref.path):
+            raise CustomError(
+                f"""Given argument resulted in invalid Firebase Realtime Database reference.
+                {ref.path}"""
+            )
         try:
             ref.delete()
         except exceptions.InvalidArgumentError:
@@ -38,7 +43,12 @@ def archive_project(project_ids: list) -> None:
                 ref.update({key: None for key in chunk})
             ref.delete()
 
-        ref = fb_db.reference(f"v2/results/{project_id}")
+        ref = fb_db.reference(f"v2/tasks/{project_id}")
+        if not re.match(r"/v2/\w+/[-a-zA-Z0-9]+", ref.path):
+            raise CustomError(
+                f"""Given argument resulted in invalid Firebase Realtime Database reference.
+                {ref.path}"""
+            )
         try:
             ref.delete()
         except exceptions.InvalidArgumentError:
@@ -49,7 +59,22 @@ def archive_project(project_ids: list) -> None:
                 ref.update({key: None for key in chunk})
             ref.delete()
 
-        fb_db.reference(f"v2/groups/{project_id}").delete()
+        ref = fb_db.reference(f"v2/groups/{project_id}")
+        if not re.match(r"/v2/\w+/[-a-zA-Z0-9]+", ref.path):
+            raise CustomError(
+                f"""Given argument resulted in invalid Firebase Realtime Database reference.
+                {ref.path}"""
+            )
+        ref.delete()
+
+        ref = fb_db.reference(f"v2/groups/{project_id}")
+        if not re.match(r"/v2/\w+/[-a-zA-Z0-9]+", ref.path):
+            raise CustomError(
+                f"""Given argument resulted in invalid Firebase Realtime Database reference.
+                {ref.path}"""
+            )
+        ref.delete()
+
         fb_db.reference(f"v2/projects/{project_id}/status").set("archived")
 
         pg_db = auth.postgresDB()
